refactoring dropdown, make it floating with popup
